Update patch set 1
Patch Set 1: Code-Review+1

Not sure if this is the correct solution, but according to JIRA comments it fixes NFC with the colorOS 2.0.x modem/fw. Tested on my find7s.

I tried updating the lib from latest kitkat release but still NFC could not be turned on.

Patch-set: 1
Label: Code-Review=+1