Update patch set 1
Patch Set 1:

NFC firmware should be downloaded to the chip from this file. If the firmware found on the NFC chip is newer than this file, no download will occur. That should be why it works without this file if you came from a stock ROM at some point.

Patch-set: 1