Update patch set 1
Patch Set 1:

The issue is, the path here has to match the path in android.  We'd need to 'fix' this whole mess by not marking the partition as sdcard in android and get it treated like a first class citizen.  I'm not sure what the best way to do that is though.

Patch-set: 1